requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytest-django==4.4.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
psycopg2>=2.8
pycodestyle==2.7.0
pycparser==2.20
pydocstyle==6.1.1
pyflakes==2.3.1
pylama==7.7.1
pylint==2.8.2
django-cors-headers==3.11.0
